Tests relocated
